Add OverloadAnOptional project with area calculation methods
Updated the solution to include a new project "OverloadAnOptional" with a static class `Geometri` for calculating the area of various geometric shapes. Added detailed XML documentation for methods. Created a project file specifying the SDK and target framework. Modified `Program.cs` to demonstrate the usage of the new methods with overloaded and optional parameters.